src/api/serializers.py
from django.contrib.auth.models import User, Group
from rest_framework import serializers
from polls.models import Question
from polls.models import Choice

class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['first_name', 'last_name', 'username', 'email', 'url', 'groups']

class GroupSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Group
        fields = ['url', 'name']


class QuestionSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model=Question
        fields=['id','question_text','pub_date','choice_set']


class ChoiceSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Choice
        fields=['choice_text','votes','question'] 

